Work around an ibrowse bug in handling HEAD requests
We were seeing ibrowse HEAD requests hang and timeout when making head
requests to AWS S3 for a key that does not exist. The issue seems to
be described by https://github.com/cmullaparthi/ibrowse/issues/67 and
we confirmed that AWS S3 is indeed setting transfer-encoding: chunked
on the responses to the HEAD queries.

Forcing the HEAD request to HTTP 1.0 works around the problem.