Use the mainline ibrowse
We've been using a very old fork of ibrowse; the mainline (cmullaparthi/ibrowse)
appears to have nearly all the things we forked it for in the first place.

Unfortunately a recent commit 555f707 (pr #155) breaks things where an
ipv6 address is reported from DNS but isn't routable. So we are
pinning to a slightly older version. This is a short term fix to get
things running while a better solution is found.

Signed-off-by: Mark Anderson <mark@chef.io>

Use locked version of ibrowse